# ssbc: Make computation of candidate queues faster