--- a/enterprise/cmd/frontend/internal/batches/resolvers/batch_spec_workspace_test.go
+++ b/enterprise/cmd/frontend/internal/batches/resolvers/batch_spec_workspace_test.go
@@ -127,6 +127,7 @@ func TestBatchSpecWorkspaceResolver(t *testing.T) {
 	t.Run("Queued", func(t *testing.T) {
 		job := &btypes.BatchSpecWorkspaceExecutionJob{
 			BatchSpecWorkspaceID: workspace.ID,
+			UserID:               userID,
 		}
 		if err := ct.CreateBatchSpecWorkspaceExecutionJob(ctx, bstore, store.ScanBatchSpecWorkspaceExecutionJob, job); err != nil {
 			t.Fatal(err)