--- a/migrations/frontend/1658174103_workspace_execution_user_queues/metadata.yaml
+++ b/migrations/frontend/1658174103_workspace_execution_user_queues/metadata.yaml
@@ -1,0 +1,2 @@
+name: workspace_execution_user_queues
+parents: [1657635365]