@eseliger 2022-07-19
> Probably foolishly mostly trusting our CI for the test updates, but the schema changes look fine. 

Because it is flaky af or because you actually are concerned this isn't well-tested enough? I can also defer this to after the cut.