@LawnGnome 2022-07-19
> Because it is flaky af or because you actually are concerned this isn't well-tested enough?

I answered this in Slack, but for the record: the former.